video_grouper/requirements.txt
httpx
aiofiles
onvif-zeep
requests
